requirements.txt
Flask
flask_cors
python-dotenv
sqlalchemy 
psycopg2-binary
voluptuous
pyjwt
cryptography
flasgger
Flask-Mail
redis
celery
openai-whisper
numpy<2